requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
idna==3.4
language-tool-python==2.7.1
packaging==23.1
Pillow==10.0.0
pytesseract==0.3.10
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.4
